+++ backend/initial/internal/zap.go
@@ -37,7 +37,7 @@ func (z *_zap) GetEncoderConfig() zapcore.EncoderConfig {
 }
 
 func (z *_zap) GetEncoderCore(l zapcore.Level, level zap.LevelEnablerFunc) zapcore.Core {
-	writer, err := FileRotatelogs.GetWriteSyncer(l.String())
+	writer, err := FileRotateLogs.GetWriteSyncer(l.String())
 	if err != nil {
 		fmt.Printf("Get Write Syncer Failed err:%v", err.Error())
 		return nil
