+++ backend/initial/migrate.go
@@ -1,9 +1,10 @@
-package cmd
+package initial
 
 import (
 	"alfred/backend/model"
 	"alfred/backend/pkg/global"
 	"fmt"
+	"github.com/spf13/cobra"
 	"os"
 )
 
@@ -42,7 +43,7 @@ func getMigrateModel() []any {
 	}
 }
 
-func migrateDB() {
+func MigrateDB(_ *cobra.Command, _ []string) {
 	if err := initSystem(); err != nil {
 		fmt.Println("init system err:", err)
 		os.Exit(1)
