+++ main.go
@@ -1,9 +1,28 @@
 package main
 
 import (
-	"alfred/backend/cmd"
+	"alfred/backend/initial"
+	"fmt"
+	"github.com/spf13/cobra"
+	"os"
 )
 
 func main() {
-	cmd.Execute()
+	rootCmd := &cobra.Command{
+		Use:   "server",
+		Short: "accounts server.",
+		Long:  `accounts server.`,
+		Run:   initial.StartServer,
+	}
+	rootCmd.AddCommand(&cobra.Command{
+		Use:   "migrate-db",
+		Short: "Auto migrate database by gorm.",
+		Long:  `Auto migrate database by gorm.`,
+		Run:   initial.MigrateDB,
+	})
+
+	if err := rootCmd.Execute(); err != nil {
+		fmt.Println("exec cmd error:", err.Error())
+		os.Exit(1)
+	}
 }
